Rate-limit the auth proxy reload and move NonJsonResponseError to shared
Review feedback: the one-shot flag was cleared on /user success, so a proxy that kept failing only /organization could reload on every navigation. Track the last reload timestamp instead and reload at most once a minute; nothing needs clearing, so recovery from a later proxy expiry works by itself. Also moves NonJsonResponseError to shared/util/errors.ts alongside PermissionError.
