+++ packages/front-end/services/auth.tsx
@@ -9,6 +9,7 @@ import React, {
 } from "react";
 import { useRouter } from "next/router";
 import { OrganizationInterface } from "shared/types/organization";
+import { NonJsonResponseError } from "shared/util";
 import {
   IdTokenResponse,
   UnauthenticatedResponse,
@@ -42,19 +43,6 @@ export function appendIgnoreWarnings(url: string): string {
 
 export function isExternalApiPath(url: string): boolean {
   return /^\/api\/v\d/.test(url);
-}
-
-// Thrown when the API answers with something other than JSON, e.g. an auth proxy (Google IAP) responding in plain text after its session expired
-export class NonJsonResponseError extends Error {
-  status: number;
-  constructor(status: number) {
-    super(
-      status === 401 || status === 403
-        ? "Your network sign-in session has expired. Reload the page to continue."
-        : `The API returned an invalid response (HTTP ${status})`,
-    );
-    this.status = status;
-  }
 }
 
 export interface AuthContextValue {
